updated jsons schema & some pipeline
PS for zeren: The task assigned for you during meeting is exempted;
It is relatively trivial for me to complete, yet can take quite a lot longer for you, as you weren’t the one designing the example json for concept node in the first place and have to re-read the entire json file to get a good understanding of it. I’ve already finished it.